pyproject.toml
[project]
name = "doc-modernizer"
version = "0.1.0"
description = "AI-powered technical documentation modernizer using LangGraph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.1.0",
    "langgraph>=0.0.40",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.0.20",
    "gradio>=4.0.0",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.11.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "lxml>=4.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
